The focus will switch to track list TableView now.
Add 'Enter' key in numpad handler . Some with 'Return' key.

Add 'Space' key response.

+++ src/PlayerEngine.mm
@@ -131,7 +131,7 @@
         
         assert(list);
         
-        int index = track.index;
+        int index = (int)track.index;
         int count = (int)[list count];
         int indexNext =-1;
         PlayOrder order = (PlayOrder)d.playOrder;
